requirements.txt
Django=2.2.12
djangorestframework=3.9.4
drf-haystack==1.8.9
psycopg2=2.8.0
